Re-factoring and re-arranging to meet norm stds
